requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
Pillow
fpdf